little fix typing

+++ handler/feeds_report.py
@@ -86,7 +86,7 @@ class FeedReport(FileMixin):
                             'category_name': '',
                             'offers_count': 0
                         }
-                    category_data[category_id]['prices'].append(int(price))
+                    category_data[category_id]['prices'].append(float(price))
                     category_data[category_id]['offers_count'] += 1
 
             def aggregate_data(category_id):
